refactor: :recycle: 유저 서비스 테스트를 통합 테스트로 리팩토링
